requirements.txt
fastapi>=0.110,<0.116
uvicorn[standard]>=0.24,<0.31
SQLAlchemy>=2.0.25,<2.1
pydantic>=2.6,<2.9
pydantic-settings>=2.2,<2.5
APScheduler>=3.10,<3.11
python-dateutil>=2.8,<2.9
httpx>=0.27,<0.28
psycopg2-binary>=2.9,<2.10
aiosqlite>=0.20,<0.21
twilio>=9.2,<10
python-dotenv>=1.0,<2
python-multipart>=0.0.9
